+++ src/app/(main)/clubs/[clubId]/comments/_components/CommentGridItem.tsx
@@ -17,15 +17,24 @@ interface CommentItemProps {
 const CommentGridItem = ({ comment, position, handleMouseDown }: CommentItemProps) => {
   return (
     <div
-      className="absolute flex flex-col items-center justify-center w-[130px] h-[170px] cursor-pointer"
+      className="absolute flex flex-col items-center justify-center w-[150px] h-[170px] cursor-pointer"
       onMouseDown={(e) => handleMouseDown(e, comment.id.toString())}
       style={{
         left: `${position.x}px`,
         top: `${position.y}px`,
       }}
     >
-      <Image src={selectPlantImg(comment.category)} alt={comment.category} width={100} height={100} priority />
-      <h3 className="w-full absolute bottom-0 text-center text-sm">{`${comment.nickname}님의 ${comment.category}`}</h3>
+      <div className="w-24">
+        <Image
+          src={selectPlantImg(comment.category)}
+          alt={comment.category}
+          width={100}
+          height={100}
+          className="object-cover"
+          priority
+        />
+      </div>
+      <h3 className="w-full absolute bottom-0 text-center text-sm">{`${comment.nickname}의 ${comment.category}`}</h3>
     </div>
   );
 };
